Refactor: change components organisation
Change components organisation :
 - Atomic components into lib,
 - Pages now have their own directory
 - Pages use boilerplate
 - molecular components are sorted by type

 Remove some duplicates (mostly CSS)
