+++ exercices/steps/apps/react/solution/src/components/cart/panel/panel.tsx
@@ -2,6 +2,7 @@ import { useContext } from 'react'
 import { Link } from 'react-router-dom'
 import { BookDTO } from '../../../api/book'
 import { CartContext } from '../../../contexts/cart-provider'
+import Header from '../../lib/header/header'
 import CartList from './list/list'
 import './panel.module.scss'
 
@@ -15,7 +16,9 @@ export function Panel ({ className }: PanelProps) {
   const { cart } = useContext(CartContext)
   return (
     <div className={className} data-cy='cart-panel'>
-      <Link to="/cart"><h2 className="pb-4 text-lg font-semibold">Cart</h2></Link>
+      <Link to="/cart">
+        <Header label="Cart" />
+      </Link>
 
       {cart?.length
         ? <CartList items={cart} />
