+++ exercices/steps/apps/react/solution/src/components/lib/input/input.tsx
