+++ exercices/steps/apps/react/solution/src/components/lib/panel/panel.tsx
