+++ exercices/steps/apps/react/solution/src/pages/cart/cart.tsx
@@ -1,12 +1,12 @@
 import './cart.module.scss'
 import { useCallback, useContext, useEffect, useState } from 'react'
-import { CartContext } from '../../../contexts/cart-provider'
-import CartList from '../../cart/panel/list/list'
-import Header from '../../header/header'
-import { UserPaymentDTO } from '../../../api/cart'
-import Address, { AddressDTO, DEFAULT_ADDRESS } from '../../user/address/address'
-import { UserContext } from '../../../contexts/user-provider'
-import CreditCard from '../../user/credit-card/credit-card'
+import { CartContext } from '../../contexts/cart-provider'
+import CartList from '../../components/cart/panel/list/list'
+import Header from '../../components/header/header'
+import { UserPaymentDTO } from '../../api/cart'
+import Address, { AddressDTO, DEFAULT_ADDRESS } from '../../components/user/address/address'
+import { UserContext } from '../../contexts/user-provider'
+import CreditCard from '../../components/user/credit-card/credit-card'
 import { useHistory } from 'react-router'
 
 /* eslint-disable-next-line */
