+++ exercices/steps/apps/react/solution/src/pages/home/home.tsx
@@ -1,0 +1,17 @@
+import GenreList from '../../components/genre/list/list'
+import Top from '../../components/books/top/top'
+import ThreeColumns from '../three-columns/three-columns'
+
+/* eslint-disable-next-line */
+export interface HomeProps {}
+
+export function Home (props: HomeProps) {
+  return (
+    <ThreeColumns>
+      <Top data-cy="top-10"/>
+      <GenreList />
+    </ThreeColumns>
+  )
+}
+
+export default Home
