+++ exercices/steps/apps/react/solution/src/pages/signin/signin.tsx
@@ -1,8 +1,8 @@
 import css from './signin.module.scss'
 import { useCallback, useContext, useState } from 'react'
 import { useHistory } from 'react-router'
-import { UserContext } from '../../../contexts/user-provider'
-import { CredentialsError } from '../../../api/user'
+import { UserContext } from '../../contexts/user-provider'
+import { CredentialsError } from '../../api/user'
 /* eslint-disable-next-line */
 export interface SigninProps {}
